fix(apisimp): guard page*pageSize overflow with (long) cast (CodeQL)
CodeQL flagged integer overflow potential in the pagination slice
arithmetic: `page` carries @PositiveOrZero but no @Max, so
`page * pageSize` can wrap for large page values.  Cast to long
before the multiply, then clamp back to int.

Fixes APISIMP-VOCAB-BROWSE-FAKE-PAGED / CodeQL finding on PR #2239.

+++ backend/src/main/java/de/dlr/shepard/v2/semantic/resources/VocabularyBrowseRest.java
@@ -151,7 +151,7 @@ public class VocabularyBrowseRest {
     List<Predicate> rows = predicateDAO.listByVocabulary(vocabId);
     List<PredicateIO> all = rows.stream().map(PredicateIO::from).toList();
     int total = all.size();
-    int from = Math.min(page * pageSize, total);
+    int from = (int) Math.min((long) page * pageSize, total);
     int to = Math.min(from + pageSize, total);
     return Response.ok(new PagedResponseIO<>(all.subList(from, to), total, page, pageSize)).build();
   }
